Added Seperator, license, publisher
