+++ package.json
@@ -14,7 +14,9 @@
     "@radix-ui/react-dropdown-menu": "^2.0.6",
     "@radix-ui/react-label": "^2.1.0",
     "@radix-ui/react-progress": "^1.1.0",
+    "@radix-ui/react-separator": "^1.1.0",
     "@radix-ui/react-slot": "^1.1.0",
+    "@radix-ui/react-tabs": "^1.1.1",
     "@tauri-apps/api": "2.0.3",
     "@tauri-apps/plugin-autostart": "2.0.0",
     "@tauri-apps/plugin-notification": "2.0.0",
