+++ src-tauri/tauri.conf.json
@@ -13,8 +13,11 @@
     "windows": [
       {
         "title": "Blink Eye",
-        "width": 400,
-        "height": 700
+        "width": 450,
+        "height": 850,
+        "center": true,
+        "maxHeight": 850,
+        "maxWidth": 450
       }
     ],
     "security": {
@@ -37,7 +40,10 @@
     "copyright": "Noman Dhoni",
     "category": "Productivity",
     "shortDescription": "A minimalist eye care reminder app.",
-    "longDescription": "",
+    "longDescription": "Blink Eye an Open Source minimalist eye care reminder app. To reduce eye strain, featuring reminder with timers, full-screen popups. Based on 20-20-20 rule.",
+    "publisher": "Noman Dhoni",
+    "license": "MIT",
+    "licenseFile": "LICENSE.txt",
     "linux": {
       "appimage": {
         "bundleMediaFramework": false,
