+++ src/components/DefaultStartMinimize.tsx
@@ -1,5 +1,5 @@
 import { invoke } from "@tauri-apps/api/core";
-import { getAllWindows } from "@tauri-apps/api/window";
+import { getAllWindows, LogicalSize } from "@tauri-apps/api/window";
 import { useEffect } from "react";
 
 const DefaultStartMinimize = () => {
@@ -21,6 +21,7 @@ const DefaultStartMinimize = () => {
         switch (result) {
           case "minimized":
             await appWindow.minimize();
+            await appWindow.setSize(new LogicalSize(0, 0));
             appWindow.onCloseRequested(async (event) => {
               event.preventDefault(); // Prevent the window from closing
               await appWindow.setSkipTaskbar(true);
