+++ src/components/backgrounds/DefaultBackground.tsx
@@ -1,0 +1,38 @@
+import React, { useEffect, useRef } from "react";
+const DefaultBackground: React.FC = () => {
+  const canvasRef = useRef<HTMLCanvasElement | null>(null);
+
+  useEffect(() => {
+    const canvas = canvasRef.current;
+    if (!canvas) return;
+
+    const ctx = canvas.getContext("2d");
+    if (!ctx) return;
+
+    const updateCanvasSize = () => {
+      canvas.width = window.innerWidth;
+      canvas.height = window.innerHeight;
+    };
+
+    updateCanvasSize();
+
+    const handleResize = () => {
+      updateCanvasSize();
+    };
+    window.addEventListener("resize", handleResize);
+
+    return () => {
+      window.removeEventListener("resize", handleResize);
+    };
+  }, []);
+
+  return (
+    <canvas
+      ref={canvasRef}
+      className="fixed inset-0 -z-10"
+      aria-hidden="true"
+    />
+  );
+};
+
+export default DefaultBackground;
